Move LabelResolver to its own file
